% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the dataframe then create and add a DATE column.}
\source{
This function calls the function eq_location_clean() within it.
}
\usage{
eq_clean_data(raw_df)
}
\arguments{
\item{This}{function takes the cleaned dataframe as inputs.}
}
\value{
This function creates a cleaned dataframe and returns a tibble.
}
\description{
This function generates a DATE column and add in to the cleaned dataframe
and it changes the class of the variables LATITUDE, LONGITUDE, EQ_PRIMARY and
DEATHS to numeric.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="NOAAA")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data()
}

}
