% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create popup text for markers for a earthquake point.}
\usage{
eq_create_label(eq_clean = NULL)
}
\arguments{
\item{cleaned}{and filtered dataframe for any location and dates.}
}
\value{
This function returns a map in HTML form with some text marker.
}
\description{
This function puts together a character string for each earthquake that will
show the cleaned location on the map.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="NOAAA")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}

}
