% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_data_read}
\alias{eq_data_read}
\title{Read the NOAA data set in a dataframe manner.}
\usage{
eq_data_read(filename)
}
\arguments{
\item{Dataset}{in a tab-delimited format.}
}
\value{
This function convert the raw input into a dataframe and returns
a tibble as an output.
}
\description{
This function reads the raw NOAA earthquake data from the base file and
converts it into a clean dataframe for further analysis.
}
\note{
If the base file does NOT exist at the specified directory, the function
throws an error message and stops executing.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="NOAAA")
eq_data_read(filename)
}

}
