% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean the location name column in the dataframe.}
\usage{
eq_location_clean(eq_cleanup)
}
\arguments{
\item{Input}{cleaned data.frame from the base file dataset.}
}
\value{
This function converts the location name column of the dataframe into
the country name and returns a tibble.
}
\description{
This function cleans the location name column and strip out the country name
in the dataframe.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="NOAAA")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_location_clean()
}

}
