% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_map}
\alias{eq_map}
\title{Create interactive map visualization of NOAA earthquake data.}
\usage{
eq_map(eq_clean = NULL, annot_col = "DATE")
}
\arguments{
\item{annot_col}{for "DATE".}

\item{The}{clean dataframe of the base NOAA dataset.}
}
\value{
This function returns an interactive map.
}
\description{
This function maps the epicenters (LATITUDE/LONGITUDE) and annotates each
point with in pop up window containing annotation data stored in a column of
the cleaned data frame of NOAA dataset.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="NOAAA")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")
}
}
