% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_time}
\alias{eq_time}
\title{Create a timeline visualization for NOAA earthquake base data.}
\source{
This function calls geom_timeline() and geom_timeline_label functions.
}
\usage{
eq_time(eq_clean = NULL, y = NULL, size = NULL, color = NULL,
  alpha = 0.4, timeline_label = FALSE, n_max = 5)
}
\arguments{
\item{size}{color alpha max y=COUNTRY}

\item{cleaned}{and filtered data.frame.}
}
\value{
This function returns a timeline visualization for a subset
of the earthquake data.
}
\description{
This function generates a timeline visualization for the NOAA
earthquake base data.  The data should be subsetted in order to create
an understandable visualization.  There are several ways to call
the function so that different visualizations are generated.  The
function is a wrapper function of all previously defined functions.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="NOAAA")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000) \%>\%
eq_time(size="EQ_PRIMARY",color="DEATHS")

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000) \%>\%
eq_time(size="EQ_PRIMARY",color="DEATHS",alpha=0.5,timeline_label=TRUE,n_max=8)

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter((COUNTRY=="USA" | COUNTRY=="CHINA") & lubridate::year(DATE) >= 2000) \%>\%
eq_time(y="COUNTRY",color="DEATHS",alpha=0.5)

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter((COUNTRY=="USA" | COUNTRY=="CHINA") & lubridate::year(DATE) >= 2000) \%>\%
eq_time(y="COUNTRY",color="DEATHS",alpha=0.5,timeline_label=TRUE)
}

}
