% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Visualize the information in the NOAA earthquakes dataset.}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{filtered}{dataframe for any country and date.}
}
\value{
This function returns layer of plot.
}
\description{
This function is a tool to visualize the information in the NOAA base file
dataset. This function plots a time line of earthquakes ranging from xmin
to xmaxdates with a point for each earthquake.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="NOAAA")
readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000) \%>\%
ggplot() +
geom_timeline(aes(x=DATE,size=EQ_PRIMARY,color=DEATHS),alpha=0.5)

readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter((COUNTRY=="USA" | COUNTRY=="CHINA") & lubridate::year(DATE) >= 2000) \%>\%
ggplot() +
geom_timeline(aes(x=DATE,y=COUNTRY,size=EQ_PRIMARY,color=DEATHS),alpha=0.5)
}

}
