% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Add annotations to the earthquake data visualization.}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n_max = 5, ...)
}
\arguments{
\item{filtered}{dataframe for any country and dates.}
}
\value{
This function returns a layer of plot.
}
\description{
This function (geom) adds a vertical line to each data point with a text
annotation (e.g. the location of the earthquake) attached to each line.
}
\examples{
\dontrun{
filename<-system.file("extdata","data.gz",package="NOAAA")
sample_USA <- readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000)

ggplot(data = sample_USA, aes(x=DATE)) + geom_timeline() +
geom_timeline_label(data=sample_USA,aes(label=LOCATION_NAME),n_max=8) + theme_classic()

ggplot() +
geom_timeline_label(data=sample_USA,aes(x=DATE,label=LOCATION_NAME))

sample_2_countries<- readr::read_delim(filename, delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter((COUNTRY=="USA" | COUNTRY=="CHINA") & lubridate::year(DATE) >= 2000)

ggplot() +
geom_timeline_label(data=sample_2_countries,aes(x=DATE,y=COUNTRY,label=LOCATION_NAME))
}

}
